[UPDATE] - Refactor Tag component structure by moving TagSkeleton to its own file, update imports in Category_Section to reflect changes, and enhance code organization for better maintainability.
