+++ src/components/common/Tag/types.ts
@@ -1,0 +1,7 @@
+interface TagProps {
+    title: string;
+    isSelected: boolean;
+    onClick: () => void;
+}
+
+export default TagProps;
