+++ src/components/index.tsx
@@ -4,7 +4,8 @@ export { default as Navbar } from "./layout/Navbar";
 export { default as ProductionCard } from "./homepage/ProductionCard";
 export { default as PromoCard } from "./homepage/PromoCard";
 export { default as Search } from "./common/Search";
-export { Tag, TagSkeleton } from "./common/Tag";
+export { default as Tag } from "./common/Tag";
+export { default as TagSkeleton } from "./common/Tag/TagSkeleton";
 export { default as HeaderSection } from "./homepage/Header_Section";
 export { default as ProductionSection } from "./homepage/Production_Section";
 export { default as PromoSection } from "./homepage/Promo_Section";
